See post on forum on Thu Mar 31, 2016: 	http://icl.cs.utk.edu/lapack-forum/viewtopic.php?f=13&t=4937
Thanks to Eugene Chereshnev (Intel)

+++ lapack/trunk/SRC/dsytrs2.f
@@ -106,7 +106,7 @@
 *>
 *> \param[out] WORK
 *> \verbatim
-*>          WORK is REAL array, dimension (N)
+*>          WORK is DOUBLE PRECISION array, dimension (N)
 *> \endverbatim
 *>
 *> \param[out] INFO
